refactor(imageitem): make to its own component
