+++ source/components/molecules/FileDisplay/FileDisplay.tsx
@@ -1,14 +1,18 @@
 import React, { useState } from "react";
 import type { NativeSyntheticEvent, NativeScrollEvent } from "react-native";
+
+import { ImageItem } from "..";
+
 import { HorizontalScrollIndicator } from "../../atoms";
-import ImageItem from "../ImageDisplay/ImageItem";
+
 import PdfItem from "../PdfDisplay/PdfItem";
+
 import { remove } from "../../../helpers/ApiRequest";
 
 import { Wrapper, Container } from "./FileDisplay.styled";
-import type { Pdf } from "../PdfDisplay/PdfDisplay";
-import type { Image } from "../ImageDisplay/ImageDisplay";
 
+import type { Pdf } from "../PdfDisplay/PdfDisplay";
+import type { Image } from "../ImageItem/ImageItem.types";
 import type { File } from "../FilePicker/FilePicker";
 
 interface Props {
