+++ source/components/molecules/ImageItem/ImageItem.tsx
@@ -1,92 +1,31 @@
 import React, { useEffect, useState } from "react";
-import PropTypes from "prop-types";
-import styled from "styled-components/native";
 import ImageZoom from "react-native-image-pan-zoom";
 import { readFile } from "react-native-fs";
 import type { GestureResponderEvent } from "react-native";
 import { TouchableOpacity, Dimensions, Image as RNImage } from "react-native";
+
 import { Icon, Button, Text } from "../../atoms";
+
 import { Modal, useModal } from "../Modal";
-import type { Image } from "./ImageDisplay";
+
 import { downloadFile } from "../../../helpers/FileUpload";
 
-const MAX_IMAGE_WIDTH = 120;
-const MAX_IMAGE_HEIGHT = 170;
+import {
+  SafeArea,
+  DefaultItem,
+  Flex,
+  DeleteBackground,
+  ButtonWrapper,
+  IconContainer,
+  ImageIcon,
+  ActivityWrapper,
+  ActivityWrapperModal,
+  ActivityIndicator,
+  MAX_IMAGE_WIDTH,
+} from "./ImageItem.styled";
 
-const SafeArea = styled.SafeAreaView`
-  flex: 1;
-`;
+import type { FileStatus, Props } from "./ImageItem.types";
 
-const DefaultItem = styled.TouchableOpacity`
-  margin-bottom: 20px;
-  margin-right: 20px;
-  padding-top: 6px;
-`;
-const Flex = styled.View`
-  flex-direction: column;
-  align-items: center;
-  padding: 0;
-  padding-top: 10px;
-`;
-
-const DeleteBackground = styled.View`
-  position: absolute;
-  top: -4px;
-  right: -12px;
-  padding: 4px;
-  elevation: 3;
-  background: #eeeeee;
-  z-index: 1;
-  border-radius: 20px;
-`;
-const ButtonWrapper = styled.View`
-  flex-direction: row;
-  margin-bottom: 40px;
-  justify-content: center;
-`;
-const IconContainer = styled.View`
-  margin: 2px;
-  elevation: 2;
-  shadow-offset: 0px 2px;
-  shadow-color: black;
-  shadow-opacity: 0.4;
-  shadow-radius: 5px;
-  border: 1px solid transparent;
-`;
-const ImageIcon = styled.Image`
-  width: ${MAX_IMAGE_WIDTH}px;
-  height: ${MAX_IMAGE_HEIGHT}px;
-`;
-const ActivityWrapper = styled.View`
-  width: ${MAX_IMAGE_WIDTH}px;
-  height: ${MAX_IMAGE_HEIGHT}px;
-  flex-direction: row;
-  justify-content: center;
-  align-items: center;
-`;
-const ActivityWrapperModal = styled.View<{ width: number; height: number }>`
-  width: ${(props) => props.width}px;
-  height: ${(props) => props.height}px;
-  flex-direction: row;
-  justify-content: center;
-  align-items: center;
-`;
-const ActivityIndicator = styled.ActivityIndicator`
-  margin-top: 12px;
-  margin-bottom: 24px;
-`;
-
-type FileStatus =
-  | "checkLocalFile"
-  | "localFileAvailable"
-  | "downloading"
-  | "downloadedFileAvailable";
-
-interface Props {
-  image: Image;
-  onRemove: () => void;
-  onChange?: (image: Image) => void;
-}
 const ImageItem: React.FC<Props> = ({ image, onRemove, onChange }) => {
   const [modalVisible, toggleModal] = useModal();
   const [fileStatus, setFileStatus] = useState<FileStatus>("checkLocalFile");
@@ -213,12 +152,6 @@ const ImageItem: React.FC<Props> = ({ image, onRemove, onChange }) => {
       </Modal>
     </React.Fragment>
   );
-};
-
-ImageItem.propTypes = {
-  image: PropTypes.object,
-  onRemove: PropTypes.func,
-  onChange: PropTypes.func,
 };
 
 export default ImageItem;
