+++ source/components/molecules/index.ts
@@ -21,3 +21,4 @@ export { default as FloatingButton } from "./FloatingButton";
 export { default as BackNavigation } from "./BackNavigation";
 export { default as ApiStatusMessage } from "./ApiStatusMessage/ApiStatusMessage";
 export { default as CloseDialog } from "./CloseDialog/CloseDialog";
+export { default as ImageItem } from "./ImageItem";
