+++ source/containers/Form/FormUploader.tsx
@@ -3,7 +3,7 @@ import { ActivityIndicator } from "react-native";
 import styled, { withTheme } from "styled-components";
 import Config from "react-native-config";
 import { Button, Heading, Icon, Text } from "../../components/atoms";
-import type { Image } from "../../components/molecules/ImageDisplay/ImageDisplay";
+import type { Image } from "../../components/molecules/ImageItem/ImageItem.types";
 import type { Pdf } from "../../components/molecules/PdfDisplay/PdfDisplay";
 import type { Options } from "../../hooks/useQueue";
 import useQueue from "../../hooks/useQueue";
